Add basic NexusUI grid sequencer
