+++ index.html
@@ -574,7 +574,7 @@
     <script type="module" src="canvasSwitcher.js"></script>
     <script type="module" src="canvasTimeline.js"></script>
     <script type="module" src="canvas-orb.js"></script>
-    <script type="module" src="pulsar.js"></script>
+    <script type="module" src="gridSequencer.js"></script>
     <script type="module" src="main.js"></script>
 </body>
 </html>
